# Allows associations to be automatically preloadable.