--- a/activerecord/lib/active_record/associations/collection_association.rb
+++ b/activerecord/lib/active_record/associations/collection_association.rb
@@ -260,6 +260,8 @@ module ActiveRecord
       end
 
       def load_target
+        automatically_preload! if automatic_preload?
+
         if find_target?
           @target = merge_target_lists(find_target, target)
         end