### Summary

Before we load an association for the first time, we can preload
the association. This will eliminate most standard N+1 queries in Rails
apps.

We do this by capturing the group of records returned by a query in an
AutomaticPreloader and having each record have a reference to it.

Once a record has been flagged to allow automatic preloading, subsequent
associations are also flagged the same so this will allow deeply nested
associations to all be dynamically preloaded

This is off by default and can be opted in with each relation. There
is also a global option to enable it by default

For example: 
This code snippet:
```ruby
developers = Developer.order(:id).limit(2).to_a
developers.each do |developer|
  developer.contracts.first.company
end
```
would typically result in a few N+1 queries.  The queries would look something like:
```sql
SELECT "developers".* FROM "developers" ORDER BY "developers"."id" ASC LIMIT 2
SELECT "contracts".* FROM "contracts" WHERE "contracts"."developer_id" = 1 ORDER BY "contracts"."id" ASC LIMIT 1
SELECT "companies".* FROM "companies" WHERE "companies"."id" = 1 LIMIT 1
SELECT "contracts".* FROM "contracts" WHERE "contracts"."developer_id" = 2 ORDER BY "contracts"."id" ASC LIMIT 1
SELECT "companies".* FROM "companies" WHERE "companies"."id" = 2 LIMIT 1
```

With `automatic_preloading` added, the same code would look like:
```sql
SELECT "developers".* FROM "developers" ORDER BY "developers"."id" ASC LIMIT 2
SELECT "contracts".* FROM "contracts" WHERE "contracts"."developer_id" IN (1, 2)
SELECT "companies".* FROM "companies" WHERE "companies"."id" IN (1, 2)
```