@nvasilevski 2022-06-01
I'm really curious about this feature! 

Correct me if I'm wrong, but I assume currently we would prevent n+1 queries by explicitly preloading associations like
```ruby
developers = Developer.includes(contracts: :company).order(:id).limit(2).to_a
```
And with the proposed change we will basically defer db queries to a later time whenever the association get's accessed without forcing developer to know in advance what kind of data they need later in the code. I can definitely see how it would prevent unintentional n+1 issues, however, and it may be my personal and a bit conservative point of view but I can see benefits in defining preloads explicitly and well in advance. Let's consider an example like:

```ruby
# developers_controller.rb
def index
  @developers = Developer.order(:id).automatic_preloading.limit(2).to_a
end
```

```erb
# developers/index.html.erb

<h1>Developers list</h1>
<% @developers.each do |developer| %>
  <div id=<%= developer.id %>>
    <h2>Contracts for developer <%developer.name%></h2>
    <ul>
    <% developer.contracts.each do |contract| %>
      <li>Contract <%=contract.title%> with company <%= contract.company.name %></li>
    <% end %>
    </ul>
  </div>
<% end %>
```

Which, as per my understanding, will basically cause the view layer to execute db queries, which doesn't generally sound like a good practice.

We could probably avoid making queries in the view layer by having some kind of a presenter that explicitly loads objects, but having a presenter is not always necessary and it would be an overkill to have a separate object that explicitly loads records. Kind of goes against the whole idea of automatic preloads. 

So what I'm trying to say, I'm just worrying that such feature will make more appealing to build such implicit designs and perhaps complicate query debugging. I personally prefer to design classes in such way to load data from datastores as early as possible and make every other part of the system almost query-less which makes it really easy to write fast unit tests that don't require to communicate with any datastore.

Again, I may be just too conservative in regards to this particular feature, but it feels like I'd still prefer explicit preloads and some separate tooling to prevent n+1 queries in the system, for example https://github.com/palkan/n_plus_one_control