@matthewd 2022-06-02
At a glance this seems to have some strong overlap with #45071, which is perhaps a good sign that it's an idea whose time has come.

Can you (cc @gwincr11) explore the similarities and differences in the proposed APIs and implementations? Don't focus on ending up with a single unified approach -- I imagine we'll have some strong opinions on the best fit for the framework -- but anything you can do to help ensure we're thinking about and considering all the right things would be great!