@doliveirakn 2022-06-02
I can see some similarities between the two approaches. maybe @gwincr11 can correct anything that I have incorrect after trying to get up to speed with that PR.

General approachs:
| | This branch | #45071 |
|-|---|---|
| Enabling point | Uses either a global flag or a method called on an ActiveRecord relation to enable it | Wraps queries in a block to enable it |
| Record storage | Stores the array of records in `ActiveRecord::Base#automatic_preloader` | Stores the array of records in `ActiveRecord::Base#_load_tree` |
 
Some things I notice:
- The block style approach does require constantly adjusting global state in order to determine if it is active. This approach can set a value in the global state initially, but otherwise only interacts with records pulled from a relation.
- With the block style approach, it would mean that query execution, and association access need to be located close to each other. (Consider the [above example](https://github.com/rails/rails/pull/45231#issuecomment-1143991277) where the initial query is executed in a controller, but associations are accessed in the view). The approach in this PR will allow those queries to not be tied to each other. 
- The block approach does give a way to explicitly disable it, whereas with this PR, once it is on for a set of records, it stays on
- The `_load_tree` concept is more generic and could theoretically be applied in other places. Knowing things like the object parent could be useful for further tooling.  However, it does appear that the `_load_tree` is always created (even if it isn't in a dynamic preloading block)
- The globally enabled aspect of this branch doesn't appear to be present with the other approach yet. I think this is really important. Turning it on globally has been a major performance improvement in the past. It also reduces developer overwhere where they no longer need to think about when to utilize it, Rails will handle it for them.