@doliveirakn 2022-06-08
So currently in this PR we have the `automatic_preloading` method that can be used both to enable or disable (very similar to how `strict_loading` works)

So someone could do `Developer.order(:id).automatic_preloading(false)` to disable it (if we was already enabled on a scope or it was globally enabled)

The only thing around disabling that is missing is for individual objects after the initial query has been made. (Theoretically they could disable it by modifying `Record#automatic_preloader` in this PR, but we could make it much more explicit.) 


Re: Memory usage. 

I don't have any information on memory usage. My argument that it wouldn't really matter that much is that most of the time, we would have loaded the same amount of objects anyways. However, there are situations, like @gwincr11 outlined, where there may be a really slow associations that we would only load on one object that would now be loaded on multiple records. Anecdotally, we have noticed no significant memory impact. 

Re: Performance. 
 
I've measured the performance of globally enabling this in two organization. Here is the changes in general response times across the latest codebase from before/after globally enabling the N+1 automatic loading. (this codebase was roughly getting about 1M requests per hour at the time)

| | Before | After |
| --- | --- | --- | 
| P50 | 55.56ms | 57.61 |  
| P75 | 101.66ms | 93.26ms | 
| P90 | 215.75ms | 166.81ms | 
| P95 | 349.52ms | 239.83ms | 
| P99 | 1.08s | 669.54ms | 
  
This data was also consistent with the previous code base I set it up on though I no longer have access to the raw numbers. The requests per hour were also well into the millions)

This to me has been really strong signal that N+1 constantly sneak through and into production, and the benefits of being able to globally remove them is really huge.